refactor: Properly scope search related state into it's own ns
