+++ src/actions/search.js
@@ -1,0 +1,4 @@
+export const setSearchFilter = searchFilter => ({
+  type: 'SET_SEARCH_FILTER',
+  searchFilter,
+});
