+++ src/reducers/index.js
@@ -4,6 +4,7 @@ import undoable, { includeAction } from 'redux-undo';
 import baseReducer from './base';
 import syncBackendReducer from './sync_backend';
 import orgReducer from './org';
+import searchReducer from './search';
 import captureReducer from './capture';
 
 const UNDOABLE_ACTIONS = [
@@ -33,5 +34,6 @@ export default combineReducers({
   org: undoable(orgReducer, {
     filter: includeAction(UNDOABLE_ACTIONS),
   }),
+  search: searchReducer,
   capture: captureReducer,
 });
